+ 1 file / commands + 1 dir / category
