+++ src/commands/fun/mod.rs
@@ -1,0 +1,6 @@
+pub mod vanish;
+pub mod trumptweet;
+pub mod phcom;
+pub mod harck;
+pub mod clyde;
+pub mod captcha;
