--- src/commands/mod.rs
@@ -1,3 +1,0 @@
-pub mod utils;
-pub mod owner;
-pub mod fun;
