+++ src/commands/owner/mod.rs
@@ -1,0 +1,3 @@
+pub mod admsay;
+pub mod quit;
+pub mod setstatus;
