+++ src/commands/owner/quit.rs
@@ -1,0 +1,23 @@
+use serenity::framework::standard::{CommandResult, macros::command};
+use serenity::model::prelude::*;
+use serenity::prelude::*;
+
+use crate::ShardManagerContainer;
+
+#[command]
+#[owners_only]
+fn quit(ctx: &mut Context, msg: &Message) -> CommandResult {
+    let data = ctx.data.read();
+
+    if let Some(manager) = data.get::<ShardManagerContainer>() {
+        manager.lock().shutdown_all();
+    } else {
+        let _ = msg.reply(&ctx, "There was a problem getting the shard manager");
+
+        return Ok(());
+    }
+
+    let _ = msg.reply(&ctx, "Shutting down!");
+
+    Ok(())
+}
