+++ src/commands/utils/mod.rs
@@ -1,0 +1,3 @@
+pub mod ping;
+pub mod say;
+pub mod links;
